v1.2

+++ database/migrations/2014_10_12_183437_create_cards_table.php
@@ -18,7 +18,7 @@ class CreateCardsTable extends Migration
             $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
             $table->string("name");
-            $table->integer("number");
+            $table->bigInteger("number");
             $table->date('expired_date')->format('m/y');
             $table->integer("cvc_number");
             $table->timestamps();
